requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
rembg==2.0.50
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0
scikit-learn==1.3.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0
lpips>=0.1.4